LA APLICACIÓN FUNCIONA JSDFLK

+++ app/src/main/java/com/app/tripup/presentation/mainFlow/itinerary/itineraryCreation/ItineraryCreationNavigation.kt
@@ -9,7 +9,7 @@ import kotlinx.serialization.Serializable
 data object ItineraryCreationDestination
 
 fun NavGraphBuilder.itineraryCreationScreen(
-    onItineraryCreated: (Long) -> Unit,
+    onItineraryCreated: (Int) -> Unit,
     onBackClick: () -> Unit
 ) {
     composable<ItineraryCreationDestination> {
